test(environment): 拆分 EnvironmentTests 测试环境类型
- 拆分 TestEnvironment 到独立文件以消除 EnvironmentTests 的 MA0048 警告

- 补充 TestEnvironment 的 XML 文档并保留原有测试行为
