+++ GFramework.Core.Tests/Events/EventBusTestsEvent.cs
@@ -1,0 +1,8 @@
+namespace GFramework.Core.Tests.Events;
+
+/// <summary>
+///     表示 <see cref="EventBusTests" /> 使用的测试事件。
+/// </summary>
+public class EventBusTestsEvent
+{
+}
